+++ fe2o3-amqp/src/transaction/owned.rs
@@ -206,7 +206,7 @@ impl OwnedTransaction {
             outcome: None,
         };
         let state = DeliveryState::TransactionalState(state);
-        let settlement = sender.inner.send_with_state(sendable, Some(state)).await?;
+        let settlement = sender.inner.send_with_state::<T, PostError>(sendable, Some(state)).await?;
 
         Ok(DeliveryFut::from(settlement))
     }